+++ tests/unit/asset/test_asset_base.py
@@ -11,7 +11,7 @@ from starfish.asset import AssetBase
 from starfish.utils.did import (
     did_to_id,
     id_to_did,
-    did_to_asset_id
+    decode_to_asset_id
 )
 
 ASSET_METADATA = {
@@ -47,14 +47,14 @@ def test_is_asset_type():
 def test_asset_id():
     asset = AssetBase(ASSET_METADATA, TEST_DID)
     assert(asset)
-    asset_id = did_to_asset_id(asset.did)
+    asset_id = decode_to_asset_id(asset.did)
     print(asset_id, asset.asset_id)
 
     assert(asset.asset_id == asset_id)
 
     did_id = secrets.token_hex(32)
     path_did = id_to_did(did_id)
-    asset_id = did_to_asset_id(f'{path_did}/{asset_id}')
+    asset_id = decode_to_asset_id(f'{path_did}/{asset_id}')
 
     asset = AssetBase(ASSET_METADATA, f'{path_did}/{asset_id}')
     assert(asset)